Added mark all read functionality
On top of recent tweets a link is added that marks all messages as read by
fading them and storing the most recent tweet in the url as an achor.

If there are new tweets when the page is reloaded the count will be reflected
on the badge of the dock icon.

The mark all read link can be accessed with ctrl+a.